requirements.txt
onnxruntime
numpy
Pillow
onnx
fastapi
uvicorn
requests